+++ custom_components/pp_reader/www/pp_reader_dashboard/js/dashboard.module.js
@@ -1,1 +1,1 @@
-export * from './dashboard.DuXhXpbP.js';
+export * from './dashboard.B-0J5Cpz.js';